Changed example and fixed directory.
